Refactoring bridging tests
