--- DentOS_Framework/DentOsTestbed/src/dent_os_testbed/test/test_suite/functional/bridging/test_bridging_wrong_fcs.py
@@ -1,122 +1,0 @@
-import pytest
-import asyncio
-
-from dent_os_testbed.lib.bridge.bridge_link import BridgeLink
-from dent_os_testbed.lib.bridge.bridge_fdb import BridgeFdb
-from dent_os_testbed.lib.ip.ip_link import IpLink
-
-from dent_os_testbed.utils.test_utils.tgen_utils import (
-    tgen_utils_get_dent_devices_with_tgen,
-    tgen_utils_get_traffic_stats,
-    tgen_utils_setup_streams,
-    tgen_utils_start_traffic,
-    tgen_utils_stop_traffic,
-    tgen_utils_get_loss,
-    tgen_utils_dev_groups_from_config,
-    tgen_utils_traffic_generator_connect,
-)
-
-pytestmark = [
-    pytest.mark.suite_functional_bridging,
-    pytest.mark.asyncio,
-    pytest.mark.usefixtures("cleanup_bridges", "cleanup_tgen")
-]
-
-async def test_bridging_wrong_fcs(testbed):
-    """
-    Test Name: test_bridging_wrong_fcs
-    Test Suite: suite_functional_bridging
-    Test Overview: Verify that packet drop due to wrong frame check sequence.
-    Test Author: Kostiantyn Stavruk
-    Test Procedure:
-    1.  Init bridge entity br0.
-    2.  Set ports swp1, swp2, swp3, swp4 master br0.
-    3.  Set bridge br0 admin state UP.
-    4.  Set entities swp1, swp2, swp3, swp4 UP state.
-    5.  Set ports swp1, swp2, swp3, swp4 learning ON.
-    6.  Set ports swp1, swp2, swp3, swp4 flood ON.
-    7.  Send traffic with bad_crc for bridge to learn address.
-    8.  Verify that addresses haven't been learned and haven't been forwarded.
-    """
-
-    bridge = "br0"
-    tgen_dev, dent_devices = await tgen_utils_get_dent_devices_with_tgen(testbed, [], 2)
-    if not tgen_dev or not dent_devices:
-        print("The testbed does not have enough dent with tgen connections")
-        return
-    dent_dev = dent_devices[0]
-    device_host_name = dent_dev.host_name
-    tg_ports = tgen_dev.links_dict[device_host_name][0]
-    ports = tgen_dev.links_dict[device_host_name][1]
-    traffic_duration = 5
-
-    out = await IpLink.add(
-        input_data=[{device_host_name: [
-            {"device": bridge, "type": "bridge"}]}])
-    assert out[0][device_host_name]["rc"] == 0, f"Verify that bridge created.\n{out}"
-
-    out = await IpLink.set(
-        input_data=[{device_host_name: [
-            {"device": bridge, "operstate": "up"}]}])
-    assert out[0][device_host_name]["rc"] == 0, f"Verify that bridge set to 'UP' state.\n{out}"
-
-    out = await IpLink.set(
-        input_data=[{device_host_name: [
-            {"device": port, "master": "br0", "operstate": "up"} for port in ports]}])
-    err_msg = f"Verify that bridge entities set to 'UP' state and links enslaved to bridge.\n{out}"
-    assert out[0][device_host_name]["rc"] == 0, err_msg
-
-    out = await BridgeLink.set(
-        input_data=[{device_host_name: [
-            {"device": port, "learning": True, "flood": True} for port in ports]}])
-    err_msg = f"Verify that entities set to learning 'ON' and flooding 'ON' state.\n{out}"
-    assert out[0][device_host_name]["rc"] == 0, err_msg
-
-    address_map = (
-        # swp port, tg port,    tg ip,     gw,        plen
-        (ports[0], tg_ports[0], "1.1.1.2", "1.1.1.1", 24),
-        (ports[1], tg_ports[1], "2.2.2.2", "2.2.2.1", 24)
-    )
-
-    dev_groups = tgen_utils_dev_groups_from_config(
-        {"ixp": port, "ip": ip, "gw": gw, "plen": plen}
-        for _, port, ip, gw, plen in address_map
-    )
-
-    await tgen_utils_traffic_generator_connect(tgen_dev, tg_ports, ports, dev_groups)
-
-    streams = {
-        "bridge": {
-            "ip_source": dev_groups[tg_ports[0]][0]["name"],
-            "ip_destination": dev_groups[tg_ports[1]][0]["name"],
-            "srcMac": "aa:bb:cc:dd:ee:11",
-            "dstMac": "aa:bb:cc:dd:ee:12",
-            "type": "raw",
-            "protocol": "802.1Q",
-            "bad_crc": True,
-        },
-    }
-
-    await tgen_utils_setup_streams(tgen_dev, config_file_name=None, streams=streams)
-
-    await tgen_utils_start_traffic(tgen_dev)
-    await asyncio.sleep(traffic_duration)
-    await tgen_utils_stop_traffic(tgen_dev)
-
-    # check the traffic stats
-    stats = await tgen_utils_get_traffic_stats(tgen_dev, "Traffic Item Statistics")
-    for row in stats.Rows:
-        assert float(row["Tx Frames"]) > 0.000, f'Failed>Ixia should transmit traffic: {row["Tx Frames"]}'
-        assert tgen_utils_get_loss(row) == 100.000, \
-            f"Verify that traffic from {row['Tx Port']} to {row['Rx Port']} not forwarded.\n{out}"
-
-    out = await BridgeFdb.show(input_data=[{device_host_name: [{"options": "-j"}]}],
-                               parse_output=True)
-    assert out[0][device_host_name]["rc"] == 0, f"Failed to get fdb entry.\n"
-
-    fdb_entries = out[0][device_host_name]["parsed_output"]
-    learned_macs = [en["mac"] for en in fdb_entries if "mac" in en]
-    list_macs = ["aa:bb:cc:dd:ee:11", "aa:bb:cc:dd:ee:12"]
-    for mac in list_macs:
-        err_msg = f"Verify that source macs have not been learned due to wrong frame check sequence.\n"
-        assert mac not in learned_macs, err_msg
